Prepare native client for multiple backends.
