+++ Cargo.toml
@@ -29,8 +29,8 @@ lazy_static = { version = "1.4", features = ["spin_no_std"] }
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 futures = "0.3"
 futures-io = "0.3.26"
-isahc = { version = "1.0", features = ["http2", "text-decoding"], default_features = false }
 uuid = { version = "1.1.2", features = ["v4"] }
+isahc = { version = "1.0", features = ["http2", "text-decoding"], default_features = false, optional = true }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 js-sys = "0.3.47"
